tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/readyState/002.html

<!doctype html>
<title>WebSockets: setting readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  ws.readyState = 5;
  assert_equals(ws.readyState, ws.CONNECTING);
});
</script>
